requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
langchain
python-dotenv
